+++ projects/ids-enterprise-ng/src/lib/sparkline/soho-sparkline.component.ts
@@ -100,6 +100,7 @@ export class SohoSparklineComponent implements AfterViewInit, AfterViewChecked, 
 
   /* Events*/
   @Output() rendered: EventEmitter<Object> = new EventEmitter<Object>();
+  @Output() contextmenu: EventEmitter<Object> = new EventEmitter<Object[]>();
 
   private jQueryElement: JQuery;
   private sparkline: SohoSparkline;
@@ -121,6 +122,8 @@ export class SohoSparklineComponent implements AfterViewInit, AfterViewChecked, 
       // Setup the events
       this.jQueryElement.on('rendered', (...args) =>
         this.ngZone.run(() => this.rendered.emit(args)));
+      this.jQueryElement.on('contextmenu', (...args) =>
+        this.ngZone.run(() => this.contextmenu.emit(args)));
     });
   }