+++ src/app/contextual-action-panel/contextual-action-panel-fullsize.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+import { SohoContextualActionPanelRef } from 'ids-enterprise-ng/lib';
+
+@Component({
+  templateUrl: './contextual-action-panel-fullsize.component.html'
+})
+
+export class ContextualActionPanelFullSizeComponent {
+  panel: SohoContextualActionPanelRef<unknown>;
+
+  close() {
+    this.panel.close();
+  }
+}